Refactored services functions/files
